Extensible context, run selection to separate file
